backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Auth
pyjwt==2.10.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# Data
pandas==2.2.3
numpy==2.2.1

# Validation & serialization
pydantic==2.10.4
pydantic-settings==2.7.1

# AI (optional, for recommendations)
anthropic==0.42.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
